Variable grid size and grid resizing
The grid is now sized based on its cell count per row (idk what to write here, you get the idea). The grid's size can be modified in the #define SIZE and it must be a float.
In the future I will handle some exceptions and maybe add some good comments/documentation.